requirements.txt
cassandra-driver==3.29.2
Flask==3.0.3
Flask-Cors==5.0.0
kafka-python==2.0.2
redis==5.2.0
vowpalwabbit==9.10.0